Gate Rust range and option patterns on source evidence
